refactor(acp): consolidate module structure and remove dead code
- Renamed modules to agent_* prefix for clarity (event_tracker → agent_event_tracker, session_flow → agent_session_flow, reconcile → agent_reconcile)
- Merged events.rs into agent_event_tracker.rs (event definitions belong with event handling)
- Merged reconcile.rs into agent_reconcile.rs (action definition + execution are cohesive)
- Removed dead code from agent.rs:
  - `get_confirmations()` / `check_approval()` (ACP doesn't track approval memory)
  - `native_skill_support()` helper (inlined at call site)
  - `supports_session_load()` / `uses_claude_meta_resume()` (moved to agent_session_flow.rs where used)
  - `require_session_id()` helper (inlined at call site)
  - Unit tests for removed methods
- Renamed public API for clarity:
  - `restore_session_id()` → `set_session_id()` (no "restore" semantics, just sets aggregate state)
  - `agent_metadata_id()` → `agent_id()` (shorter, clearer)
- Simplified imports (removed unused Confirmation, SetSessionModeRequest/SetSessionModelRequest from agent.rs public surface)
- Updated justfile: `lint-fix` now runs `cargo fix` (clippy --fix was breaking staged changes)
- Updated justfile: `push` uses recipe dependencies instead of inline commands

Why: After the aggregate refactor, several helpers lived in the wrong module (query methods far from their only call site), and some code became dead (pre-aggregate permission tracking stubs). This consolidates related code, removes the dead weight, and makes the module boundaries match actual responsibilities.
