+++ crates/aionui-ai-agent/src/manager/acp/agent_reconcile.rs
@@ -1,0 +1,133 @@
+use crate::manager::acp::AcpAgentManager;
+
+use crate::manager::acp::mode_normalize::normalize_requested_mode;
+use crate::shared_kernel::{ConfigKey, ConfigValue, ModeId, ModelId};
+use agent_client_protocol::schema::{
+    SessionId, SetSessionConfigOptionRequest, SetSessionModeRequest, SetSessionModelRequest,
+};
+use tracing::{error, info};
+
+/// Actions the session driver must execute to align CLI state with user intent.
+///
+/// Produced by `AcpSession::plan_reconcile` — a pure function that compares
+/// desired vs observed and returns a list of idempotent, order-independent ops.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum ReconcileAction {
+    SetMode { mode: ModeId },
+    SetModel { model: ModelId },
+    SetConfigOption { key: ConfigKey, value: ConfigValue },
+}
+
+impl AcpAgentManager {
+    /// Execute reconcile actions produced by `AcpSession::plan_reconcile`.
+    ///
+    /// Compares the aggregate's desired state against what the CLI has
+    /// reported as current, then issues the minimal set of SDK calls
+    /// (set_mode, set_config_option) to bring the CLI into alignment.
+    /// Best-effort: individual failures are logged but do not abort.
+    pub(super) async fn reconcile_session(&self, session_id: &str) {
+        use crate::manager::acp::ReconcileAction;
+
+        let actions = {
+            let session = self.session.read().await;
+            session.plan_reconcile()
+        };
+
+        for action in actions {
+            match action {
+                ReconcileAction::SetMode { mode } => {
+                    let normalized = normalize_requested_mode(&self.params.metadata, mode.as_str());
+                    if normalized.is_empty() {
+                        continue;
+                    }
+                    if let Err(e) = self
+                        .protocol
+                        .set_mode(SetSessionModeRequest::new(
+                            SessionId::new(session_id),
+                            normalized.clone(),
+                        ))
+                        .await
+                    {
+                        error!(
+                            conversation_id = %self.params.conversation_id,
+                            mode_id = %normalized,
+                            error = %e,
+                            "reconcile_session: set_mode failed"
+                        );
+                        continue;
+                    }
+                    // SDK does not push a notification after a successful
+                    // set_mode — sync observed/advertised ourselves so the
+                    // next plan_reconcile is a no-op.
+                    let mut session = self.session.write().await;
+                    session.apply_observed_mode(ModeId::new(normalized));
+                    self.commit_session_changes(&mut session).await;
+                }
+                ReconcileAction::SetModel { model } => {
+                    if let Err(e) = self
+                        .protocol
+                        .set_model(SetSessionModelRequest::new(
+                            SessionId::new(session_id),
+                            model.as_str().to_owned(),
+                        ))
+                        .await
+                    {
+                        error!(
+                            conversation_id = %self.params.conversation_id,
+                            model_id = %model,
+                            error = %e,
+                            "reconcile_session: set_model failed"
+                        );
+                        continue;
+                    }
+                    // SDK does not push a CurrentModelUpdate notification —
+                    // sync observed/advertised ourselves.
+                    let mut session = self.session.write().await;
+                    session.apply_observed_model(model);
+                    self.commit_session_changes(&mut session).await;
+                }
+                ReconcileAction::SetConfigOption { key, value } => {
+                    if let Err(err) = self
+                        .protocol
+                        .set_config_option(SetSessionConfigOptionRequest::new(
+                            SessionId::new(session_id),
+                            key.as_str().to_owned(),
+                            value.as_str().to_owned(),
+                        ))
+                        .await
+                    {
+                        info!(
+                            config_id = %key,
+                            desired = %value,
+                            error = %err,
+                            "reconcile_session: set_config_option failed; skipping"
+                        );
+                        continue;
+                    }
+                    // Sync observed ourselves so the next plan_reconcile
+                    // does not replay this action. CLI does not push a
+                    // config-update notification after set_config_option.
+                    let mut session = self.session.write().await;
+                    session.apply_observed_config(key, value);
+                    self.commit_session_changes(&mut session).await;
+                }
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn reconcile_action_equality() {
+        let a = ReconcileAction::SetMode {
+            mode: ModeId::new("plan"),
+        };
+        let b = ReconcileAction::SetMode {
+            mode: ModeId::new("plan"),
+        };
+        assert_eq!(a, b);
+    }
+}
