+++ crates/aionui-ai-agent/src/manager/acp/agent_session_flow.rs
@@ -1,4 +1,5 @@
 use crate::manager::acp::AcpAgentManager;
+use crate::manager::acp::mode_normalize::agent_metadata_uses_claude_meta_resume;
 use crate::protocol::events::{
     AgentStreamEvent, AvailableCommandsEventData, FinishEventData, SessionAssignedEventData, StartEventData,
 };
@@ -59,7 +60,7 @@ impl AcpAgentManager {
     /// - Backends that support neither: seed the aggregate and hope the CLI
     ///   still recognises the id (legacy behaviour — matches pre-refactor).
     pub(super) async fn open_session_resume(&self, session_id: &str) -> Result<String, AppError> {
-        if self.uses_claude_meta_resume() {
+        if agent_metadata_uses_claude_meta_resume(&self.params.metadata) {
             let mut meta = serde_json::Map::new();
             let mut claude_code = serde_json::Map::new();
             let mut options = serde_json::Map::new();
@@ -235,6 +236,17 @@ impl AcpAgentManager {
                     commands: cmds.to_vec(),
                 }));
         }
+    }
+
+    fn supports_session_load(&self) -> bool {
+        self.params
+            .metadata
+            .handshake
+            .agent_capabilities
+            .as_ref()
+            .and_then(|caps: &Value| caps.get("load_session"))
+            .and_then(|v: &Value| v.as_bool())
+            .unwrap_or(false)
     }
 }
