+++ crates/aionui-ai-agent/src/manager/acp/mod.rs
@@ -1,16 +1,15 @@
 pub mod agent;
+pub mod agent_event_tracker;
+pub mod agent_reconcile;
+mod agent_session_flow;
 pub mod catalog_forwarder;
-mod event_tracker;
-pub mod events;
 mod mode_normalize;
 pub mod permission_router;
-pub mod reconcile;
 pub mod session;
-mod session_flow;
 
 pub use agent::AcpAgentManager;
+pub use agent_event_tracker::AcpSessionEvent;
+pub use agent_reconcile::ReconcileAction;
 pub use catalog_forwarder::CatalogForwarder;
-pub use events::AcpSessionEvent;
 pub use permission_router::PermissionRouter;
-pub use reconcile::ReconcileAction;
 pub use session::AcpSession;
